djangoProject/requirements.txt
asgiref==3.5.2
attrs==21.4.0
backports.zoneinfo==0.2.1
certifi==2022.6.15
charset-normalizer==2.1.0
coreschema==0.0.4
Django==4.0.6
djangorestframework==3.13.1
idna==3.3
importlib-resources==5.8.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.7.2
MarkupSafe==2.1.1
packaging==21.3
pyparsing==3.0.9
pyrsistent==0.18.1
pytz==2022.1
PyYAML==6.0
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
sqlparse==0.4.2
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.10
zipp==3.8.1
